OvmfPkg: Introduce NetworkCfgLib
Introduce NetworkCfgLib which will set PcdNetworkSupport based on
"etc/networking" qemu file.

If "etc/networking" (type bool) is TRUE, then PcdNetworkSupport will be
TRUE and vice versa.

In the following patch, PcdNetworkSupport will be used to enable/disable
VirtIo net driver so that VMM will have control over networking
functionality in runtime.

The default value of PcdNetworkSupport is TRUE, which means if network
support is turned on at compile time and VMM doesn't do anything, the
VirtIo driver will be enabled. This is to make it consistent with the
behavior before this patch.

Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Cc: Anthony Perard <anthony.perard@citrix.com>
Cc: Julien Grall <julien@xen.org>

Signed-off-by: Yuan Yu <yuanyu@google.com>