MdeModulePkg: Fix bug in ScsiBusDxe/ScsiBus.c
A while loop in SCSIBusDriverBindingStart() is supposed to scan all the
possible Puns in the SCSI channel by calling ScsiScanCreateDevice() for
each of them. Therefore, we should not abort the loop even when one of
the Puns is disconnected.

The following is one of the scenarios.

SCSIBusDriverBindingStart()
> ScsiScanCreateDevice()
  > DiscoverScsiDevice()
    > ScsiInquiryCommand()
      > ...
        > ParseResponse()

When virtio-scsi returns VIRTIO_SCSI_S_BAD_TARGET, ParseResponse() in
VirtioScsi.c will return EFI_TIMEOUT:

    case VIRTIO_SCSI_S_BAD_TARGET:
      //
      // This is non-intuitive but explicitly required by the
      // EFI_EXT_SCSI_PASS_THRU_PROTOCOL.PassThru() specification for
      // disconnected (but otherwise valid) target / LUN addresses.
      //
      Packet->HostAdapterStatus =
        EFI_EXT_SCSI_STATUS_HOST_ADAPTER_TIMEOUT_COMMAND;
      return EFI_TIMEOUT;

This will eventually cause DiscoverScsiDevice() to return FALSE, which
will cause ScsiScanCreateDevice() to return EFI_OUT_OF_RESOURCES:

  if (!DiscoverScsiDevice (ScsiIoDevice)) {
    Status = EFI_OUT_OF_RESOURCES;
    goto ErrorExit;
  }

In sum, "disconnected (but otherwise valid) target / LUN addresses" can
result in EFI_OUT_OF_RESOURCES and when this happens the while loop in
SCSIBusDriverBindingStart() should continue, not abort.

Without this fix, the loop can terminate prematurely with good devices
not having a chance to be discovered. If this good device is the boot
device, boot will fail.

Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Hao A Wu <hao.a.wu@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Sivaparvathi chellaiah <sivaparvathic@ami.com>

Signed-off-by: Yuan Yu <yuanyu@google.com>